(feat) added env. file and config file to use API URL to easier change URL for Azure or localhost. (modified) checklist component, clean up of code for better (fix) API URLS to context and hooks (fix) overall refactor of code to improve readability and performance (fix) removal of unused imports and variables (feat) UserChip component to display user information based on user role in checklist workflow (feat) ad ChipStatus component to display status of checklist workflow

+++ src/pages/users/context/userContextProvider.tsx
@@ -8,13 +8,13 @@ import React, {
     useState,
 } from 'react'
 import { useLocation } from 'react-router'
+import { API_URL } from '../../../config'
 import useAuth from '../../landingPage/context/LandingPageContextProvider'
 import { useAddUser } from '../addUser/hooks/useAddUser'
 import { Option } from '../context/models/OptionsEntity'
 import { AzureUserInfo } from './models/AzureUserEntity'
 import { UserEntity } from './models/UserEntity'
 import { UserListEntity } from './models/UserListEntity'
-
 export type ContextType = {
     result: UserEntity[]
     userList: UserListEntity[]
@@ -72,14 +72,15 @@ const UserContextProvider = ({ children }: { children: React.ReactNode }) => {
     const { user } = useAddUser()
 
     const getUsers = async () => {
-        const res = await fetch('https://turbinsikker-api-lin-prod.azurewebsites.net/Api/GetAllUsersAdmin',
-        {
-            method: "GET",
+        if (!accessToken) return
+        const res = await fetch(`${API_URL}/GetAllUsersAdmin `, {
+            method: 'GET',
             headers: {
-              Authorization: `Bearer ${accessToken}`,
-              "Content-Type": "application/json",
-              "Access-Control-Allow-Origin": '*'
-            }})
+                Authorization: `Bearer ${accessToken}`,
+                'Content-Type': 'application/json',
+                'Access-Control-Allow-Origin': '*',
+            },
+        })
         if (!res.ok) throw new Error('Failed with HTTP code ' + res.status)
         const data = await res.json()
         setResult(data)
@@ -100,16 +101,15 @@ const UserContextProvider = ({ children }: { children: React.ReactNode }) => {
 
     useEffect(() => {
         const fetchUserRoles = async () => {
-            const res = await fetch(
-                'https://turbinsikker-api-lin-prod.azurewebsites.net/api/GetAllUserRoles',
-                {
-                    method: "GET",
-                    headers: {
-                      Authorization: `Bearer ${accessToken}`,
-                      "Content-Type": "application/json",
-                      "Access-Control-Allow-Origin": '*'
-                    }}
-            )
+            if (!accessToken) return
+            const res = await fetch(`${API_URL}/GetAllUserRoles  `, {
+                method: 'GET',
+                headers: {
+                    Authorization: `Bearer ${accessToken}`,
+                    'Content-Type': 'application/json',
+                    'Access-Control-Allow-Origin': '*',
+                },
+            })
             if (!res.ok) throw new Error('Failed with HTTP code ' + res.status)
             const data = await res.json()
 
@@ -126,13 +126,14 @@ const UserContextProvider = ({ children }: { children: React.ReactNode }) => {
     // Delete user //
 
     const handleDeleteUser = async (id: string | undefined) => {
-        await fetch(`https://turbinsikker-api-lin-prod.azurewebsites.net/api/SoftDeleteUser?id=${id}`, {
+        if (!accessToken) return
+        await fetch(`${API_URL}/SoftDeleteUser?id=${id}  `, {
             method: 'DELETE',
             headers: {
                 Authorization: `Bearer ${accessToken}`,
-                "Content-Type": "application/json",
-                "Access-Control-Allow-Origin": '*'
-              }
+                'Content-Type': 'application/json',
+                'Access-Control-Allow-Origin': '*',
+            },
         })
 
         setRefreshUsers((prevRefresh) => !prevRefresh)
@@ -161,14 +162,15 @@ const UserContextProvider = ({ children }: { children: React.ReactNode }) => {
     }
     async function fetchUserByEmail(userEmail: string, name: string) {
         const response = await fetch(
-            `https://turbinsikker-api-lin-prod.azurewebsites.net/Api/GetUserByAzureAdUserId?azureAdUserId=${userEmail}`,
+            `${API_URL}/GetUserByAzureAdUserId?azureAdUserId=${userEmail}`,
             {
-                method: "GET",
+                method: 'GET',
                 headers: {
-                  Authorization: `Bearer ${accessToken}`,
-                  "Content-Type": "application/json",
-                  "Access-Control-Allow-Origin": '*'
-                }}
+                    Authorization: `Bearer ${accessToken}`,
+                    'Content-Type': 'application/json',
+                    'Access-Control-Allow-Origin': '*',
+                },
+            }
         )
         if (response.ok) {
             const user = await response.json()
@@ -197,25 +199,22 @@ const UserContextProvider = ({ children }: { children: React.ReactNode }) => {
         const lastName = name.split(' ')[1]
         const username = `${firstName.toLowerCase()}.${lastName.toLowerCase()}`
         try {
-            const createUserResponse = await fetch(
-                'https://turbinsikker-api-lin-prod.azurewebsites.net/Api/addUser',
-                {
-                    method: 'POST',
-                    headers: {
-                        'Content-Type': 'application/json',
-                        Authorization: `Bearer ${accessToken}`,
-                        "Access-Control-Allow-Origin": '*'
-                    },
-                    body: JSON.stringify({
-                        azureAdUserId: userEmail,
-                        firstName: firstName,
-                        lastName: lastName,
-                        userName: username,
-                        email: userEmail,
-                        // other user properties
-                    }),
-                }
-            )
+            const createUserResponse = await fetch(`${API_URL}/addUser`, {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                    Authorization: `Bearer ${accessToken}`,
+                    'Access-Control-Allow-Origin': '*',
+                },
+                body: JSON.stringify({
+                    azureAdUserId: userEmail,
+                    firstName: firstName,
+                    lastName: lastName,
+                    userName: username,
+                    email: userEmail,
+                    // other user properties
+                }),
+            })
             console.log(
                 'User creation response status:',
                 createUserResponse.status
